Breaking: getBalance modified + broadcasted event improvements (dashevo/wallet-lib#59)
* refactor(insight): remove all insight reference

* chore(package): bump version to 3.0.1

Breaking change being the drop of support of Insight happened in v.3.0 but we forgot some elements.

* refactor(keychain): move console.error to throwed error

* test(keychain): improve keychain tests

* fix(keychain): generateKeyForChild from HDPubKey failing

* test(getBalance): bring getBalance test

* test(getTransaction): added sanity test

* fix: return all addresses matching the tx

* fix: wrong input consideration + excessive event

So there is two thing here : First the inputs sections was returning us the wrong amount of satoshis in some context. Also the UNCONFIRMED_BALANCE_CHANGED event should send : -sumSpent+sumSpentToSelf, but actually we will delegate this event to other part.

* breaking: change BALANCE_CHANGED to CONFIRMED_BALANCE_CHANGED

This is done to avoid any type of confusion

* feat: Storage.calculateDuffBalance

Added ability to perform calculation of duff balance of specific walletId/Index. Is needed to allow storage.updateAddress to know balance for events. See next commit

* breaking: getBalance separated in three method : getTotal, getUnconf,..

This is due to avoid confusion. Also usage of new storage method instead of calculating with access from addresses storage.

* fix: tests for getBalance

* refact: removed unnecessary calculation

* fix: erroneous parameters in getBalance methods. Removed first param.

* typo: updated debug text balance -> confirmed balance